a2a_real_llm/host_agent_adk/pyproject.toml
[project]
name = "host_agent_adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.8",
    "dotenv>=0.9.9",
    "google-adk>=1.3.0",
    "httpx>=0.28.1",
    "litellm>=1.72.6",
    "nest-asyncio>=1.6.0",
    "ollama>=0.5.1",
    "openai>=1.88.0",
    "uvicorn>=0.34.3",
    "yfinance>=0.2.63",
]
